requirements.txt
cassandra-driver
pymongo
pydgraph
python-dotenv